VHE: Update vm helper functions to account for EL0 partitions
Updated the VM helper functions that init and map memory into the VM to
account for EL0 partitions. When a partition is an el0 partition, we use
the appropriate stage 1 mapping API's to ensure they are mapped as stage
1 tables and entries.
Note that it is weird to call vm helper functions for an EL0 entity but
a EL0 partition is being modeled as a lightweight VM to promote
reusability of code and not have to create a parallel universe of
helpers, functions and management.

Change-Id: Iafb4fdd3e921edff4d523c9a9fb159ddb9f4bb7f
Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>