VHE: Switch to hypervisor/host translation tables on entry to EL2
On every entry into the hypervisor, change ttbr0_el2 to point to the
ASID 0 hypervisor tables. All of hypervisor code, except for the few
instructions during entry, will run on the hypervisor tables. This is
done so that hypervisor code with and without VHE are identical in that
they always use the same page tables during execution. Also, EL0 pages
are not mapped in the hypervisor for security and hygiene.

Change-Id: Ia5e3e8eddad29e325d1694d2ed3d2e39286f5589
Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>