VHE: Tests: Add FF-A test suite to EL0 partitions
Add the FF-A test suite from primary_with_secondaries tests to run EL0
partitions. All the tests should work as is even on EL0 partitions since
the same rules for message passing apply to EL0 partitions as it does
for VMs. This patch also adds a couple of service VMs as well, as
required by the test suite. The service VMs dont do anything yet.

Change-Id: I2fa9ecf18e798d7d00e8e19bcf33011c87b6cdfb
Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>