VHE: Tests: Add mailbox tests to EL0 partitions.
This patch adds the mailbox test suite from the primary_with_secondaries
test suite and applies it to EL0 partitions. Most of the tests remain
exactly the same, except the primary_to_secondary test. This particular
test injects interrupts into a VM, but since interrupt injection is not
allowed, we need to remove the code. The error code expected when the
partition uses a WFE or WFI is different for EL0 partitions as well,
since the partition just yields on use of WFE/WFI. The
echo_with_notification service is also modified to remove
interrupt/exception handling and any other dependencies on interrupts
since those dont apply to EL0 partitions.

Change-Id: I1a82990ed5abdcfc3f5e2e045cbbc3abf1563d49
Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>