VHE: Tests: Add interrupt tests for EL0 partitions
Added interrupt test suite from primary_with_secondaries test suite and
applied it to EL0 partitions. The main difference between a VM and EL0
partition is that there is no irq handler in an EL0 partition. However,
the FFA_INTERRUPT status interface can be used to notify EL0 partitions
that there are interrupts pending, at which the irq handlers are called
manually as opposed to architecturally by the CPU itself.
Also, the wfi test is removed since an interrupt inject from the primary
VM really just sets a bit and not inject interrupts through hcr_el2
since it doesnt apply.

Change-Id: I78b786ab41fc8f03d7a5974807da472a8cdc226c
Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>