Fix Tests: Initialization of console stalls tests
Map the system address space from address 0x1000 such that the mm
library creates a L2 table to which the first L1 descriptor points to.
Provided SPMC and SP images reside below 1GB, same as peripherals this
is to prevent the case that the mm library has to break down the first
L1 block descriptor while it is currently executing from a region mapped
by the same L1 descriptor which is not architecturally possible.
Update to TF-A prebuilt 'bl31_spmd.bin', with following commit:
2e0e51f42586826a1f6f6c1e532f90e6df642cf5.

Change-Id: I9a9ff6e8fc6714fe952d333f0146887d064ce1f6
Signed-off-by: J-Alves <joao.alves@arm.com>
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>