Bug Fix: Fix memory relend test
The relend_after_return memory test has a bug where it lends memory to
the secondary VM twice and expects the secondary VM to access it, and
then NOT fault. However, the secondary VM runs the
ffa_memory_lend_relinquish function that does not retrieve the second
memory lend request, causing it to fault. The secondary VM function
registers an exception handler, which uses FFA_MSG_SEND to send the
exception count back to the primary VM, and the primary VM does not
check to ensure the exception count is 0. The test logs clearly show the
secondary VM faulting but the test still passes.
This patch fixes the issue by changing the primary VM to ensure that the
arguments for the message send from the secondary VM is 0 (since
exception handlers return the count in the rx/tx buffers with count of
4). The secondary VM code is updated to add a
ffa_memory_lend_relinquish_relend function that enters a loop where it
accepts and relinquishes memory requests.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@icloud.com>
Change-Id: I007e12cea8ca70f6094ccce3a54db42fa988a6a8