feat: FF-A mem share pass security state attribute
Update FF-A memory send functions to track the security state of a
memory region into the internal share states. This permits delivering
the right NS bit state to the FF-A mem retrieve response per FF-A v1.1
REL0 section 10.10.4.1, and perform the appropriate mapping and
maintenance depending on the targetted NS/S IPA range.

Change-Id: I70cf1ed8444346325877a0b4922be81662036899
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>