Tests: Decouple Hypervisor and SPMC test drivers
Going forward it is intended to support a test set-up that encompasses
the Hypervisor and the SPMC. This patch intends to facilitate the
support of such test set-up. Added arguments to the script that are
specific to Hypervisor or SPMC driver, and slightly refactored the
FVP driver so that they can be used simultaneously.

Change-Id: Ic83f09bcb873d4cdea88ec1d1a7268d322816ba3
Signed-off-by: J-Alves <joao.alves@arm.com>