Tests: Reconcile partition's manifest extensions
The Partitions Manager(PM) manifest is expected to contain information
that is partition-specific, within node "hypervisor".
The hftest.py is in charge of appending the partitions nodes the PM
manifest, build it and load it into the platform. This patch
improves that process by autonomously changing the nodes name to comply
with expected naming sequence, making it easier to reutilize the nodes
across different test set-ups with various partitions combinations.

Change-Id: Icd6c34862aaabf7b7e3f5922091e28fc264c9c05
Signed-off-by: J-Alves <joao.alves@arm.com>