VHE: Tests: Secure EL0 partitions test
So far, VHE support and EL0 tests have been run on the normal world.
This patch adds test support to be able to run S-EL0 partitions and test
them just like EL0 partitions in the normal world. Specifically, the
test setup contains a primary partition in the normal world that drives
the tests, one S-EL1 partition in the secure world, that runs as the
equivalent of the primary partition in the secure world (required by
secure world implementation due to PSCI limitations) and a S-EL0
partition that runs a message loop to run tests.

Change-Id: Ieddd11e0d52ec9b69eb8ee87f326d11bc3697d4c
Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>