Tests: Fix exception level in ffa_secure_partitions manifest
The normal world primary VM's exception level was marked as S-EL1
instead of EL1. This change fixes the manifest to mark it as an EL1
partition.
This change also renames the file to use underscores(_) instead of
dashes(-) for the file name, keeping it consistent with other files in
the directory.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: If300e863b998343a9e9331ed1a064b1c16a64bfb