fix(ff-a): check parameters of FFA_MSG_WAIT
Calling FFA_MSG_WAIT with non-zero values should return
INVALID_PARAMETERS as far as SPM is concerned. Usage of these
parameters for passing timeout values through ERET is out of scope as
it only concerns Hypervisor.
In addition to the check the Hafnium test has been added to check for
that.

Signed-off-by: Maksims Svecovs <maksims.svecovs@arm.com>
Change-Id: Ia223de947ea5d24bdc71a09a3cd733a28485c050