feat(ffa): secure interrupt management in SWd
This patch adds support for secure interrupt management while running
in secure world.

The secure interrupt is delegated to target secure partition where the
interrupt gets handled appropriately.

As detailed in the section titled "Secure Interrupt Signaling mechanism"
in the FF-A v1.1 spec, we leverage ERET and vIRQ conduits for interrupt
delegation.

In the current iteration, we only support S-EL1 partitions.

Change-Id: Ifd4d0ca3fec8f2ffef4c2d695b073a536aa97a28
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>