test(VHE): Add tests for FFA_MEM_PERM_GET/SET
This patch adds tests for the aforementioned ABIs. It also updates the
EL0 entry function to account for the fact that hafnium marks all of EL0
partition memory as RX during load/init. The partition must use its own
methods to determine layout and use the ABI's to set permissions for the
image appropriately.

Change-Id: If039224969a0dde3b2af78cde3c13e89ee828f4a
Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>