fix(clang): remove cognitive complexity tests
With clang 12 it is not possible to ignore the weighting added
to the cognitive complexity score of functions by macros. Since
we read macro usage like function calls this causes a lot of failures
that aren't of concern so until we update to clang 13 or higher where
we can ignore macros disable the cognitive complexity tests.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I4d44c83185ef45d3419904f7f3b11463fbd4772d