build: use a toolchain different from prebuilts
This change improves the project build scalability by:
- Providing the clang toolchain path through the PATH environment
variable before invoking make. It means there is no longer a strict
dependency to the toolchain stored in prebuilts. The latter is still
required/used in context of the CI.
- Updating prebuilts tree hash to provide tools required by Arm host
native builds.
- ninja/gn tools path can be provided through NINJA and GN variables on
the make command line (no change). dtc tool is found from PATH.
It can be specified to point to the default prebuilt binary.

Change-Id: Iefb14564554f7d907751ba82db6884fe7fa5c50c
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>