refactor(ci): simplify SPMC test setups
To test Hafnium as SPMC, we supported two different test setups:
1- SPMC and SPMD, with two SPs;
2- Hypervisor and SPMC, with a VM controlling an SP.

The build artifacts to obtain the SPs for test setup 1 are deleted in
this patch. The intention is to simplify the code structure of the
tests.
Subsequent changes should focus on improving aforementioned setup 2.
For tests that need to validate SP-to-SP functionality, the test
should always be initialized from the VM, which should communicate via
direct messaging with an SP to, in its turn, test the necessary
functionality.

Change-Id: I45ec82fa17b0dcb138d0e562ab57f0afb2e07939
Signed-off-by: J-Alves <joao.alves@arm.com>