refactor(ci): move dir msg test to standalone SP
Test introduced in change [1], to validate that SP can't use
FFA_MSG_SEND_DIRECT_REQ to send a message to the NWd was moved to
to standalone SP image.
Doing this as file "partition_control.c" where the test was defined is
now intended to be used as a VM only.

[1] I097547f48683b9bf2da86d5304f62453742c974b

Change-Id: I3ca14797a8fce40268b1015ef1575a804fa385db
Signed-off-by: J-Alves <joao.alves@arm.com>