refactor(ci): ffa_secure_partitions test setup
The test setup is being refactored to simplify the code structure:
- The file in "partition_control.c" is now intended to be deployed as a
VM, as such removed dependency with SECURE_WORLD macro.
- Moved "partition_services.h" into the scope of ffa_secure_partitions.

With this change, tests are now supposed to be controlled by this VM,
which can:
- Perform FF-A calls directly to the SPMC.
- Send direct msg request to the deployed SPs (currently there is only
one), with test-specific payloads that SPs should use to determine the
test code to execute.

Change-Id: I0e182fbd6327f0a08dfc4e69311aeb201beedbd4
Signed-off-by: J-Alves <joao.alves@arm.com>