fix(clang): fix performance-no-int-to-ptr for tests
To keep inline with table 10.14 of the FF-A 1.1 BETA0 specification
"Constituent memory region descriptor" we keep the address field
in the ffa_memory_region_constituent struct as a uint64_t to ensure
it is always 8 bytes. This means we will have to have integer to pointer
conversions but for the cases of this in the buffer manipulation tests
we will permit this.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I6688af1c8cd3d959e0808f7568f190218d8aa52d