fix(ff-a): plat_ffa module build failures
The patch corrects the following issues with plat_ffa module:
- The header "ffa.h" from plat_ffa module was not being included in the
source files that implement it (spmc.c, hypervisor.c and ffa.c).
This was letting some functions to link with mis-aligned signatures from
the header and source file.
- The toolchain "android_aarch64" missed plat_ffa module. This patch
includes the hash for the submodule "project/reference" that is fixing
this.
- Let "make tidy" add changes, prevent others.

Change-Id: I3e7d9c11dd26853ed1b8bbb6b6eba3f6252046b3
Signed-off-by: J-Alves <joao.alves@arm.com>