refactor(notifications): improve notifications state handling
After the moment a notification is set, the receiver scheduler should
get an indication that the notification is pending.
For this the SPMC triggers the SRI, that should be handled by the FF-A
driver in the NWd, which shall call FFA_NOTIFICATION_INFO_GET and
notify the receiver scheduler that the receiver needs CPU cycles.
The SRI is being handled in a way such that spurious triggering of the
SRI is mitigated. A spurious SRI would be one such that after a call
to FFA_NOTIFICATION_INFO_GET there would be no information returned.
The mitigation is implemented through a state machine for the SRI.
If a call to FFA_NOTIFICATION_INFO_GET didn't return all the information
of pending notifications, the FF-A Beta0 spec mandates that this is
indicated to the receiver scheduler through the return to
FFA_NOTIFICATION_INFO_GET call.
In a system with concurrent calls to FFA_NOTIFICATION_INFO_GET and
FFA_NOTIFICATION_SET, the current implementation could allow for
a case in which a notification is pending, and its state hasn't been
indicated to the receiver scheduler: neither through triggering the SRI,
nor through the return to FFA_NOTIFICATION_INFO_GET.

This patch introduces global counters to the state of notifications:
1- Counter to the number of notifications pending - incremented at
notification set, and decremented at notification get.
2- Counter to the number of notifications that have been retrieved by
receiver scheduler that are still pending - incremented at
notification info get, and decremented at notification get if the
notification has been retrieved by the receiver scheduler.

Change-Id: Icd2bdab7c825ba607c09f3cf54c5b413e7264c1e
Signed-off-by: J-Alves <joao.alves@arm.com>