fix(vmlib): Fix unaligned structure copy
ffa_memory_region_constituents use direct structure copy and does not
use memcpy. In case the pointer to the constituent structure is
unaligned, which is quiet often the case when the memory region
descriptors are being constructed, the structure copy can fail due to
alignment faults. This especially shows up when running EL0 host
applications with VHE, since SCTLR_A bit is set, causing alignment
faults. In the case of test VMs, alignment checking is not enabled by
default due to which the issue is not seen with current tests.
memcpy or memcpy_s is not used and a new helper function is created
since hafnium does not have memcpy, and the linux kernel does not have
memcpy_s.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: Ic5a0cc4c54a7c5b6d877b29cb255caf31f43f348