fix(bti): Add branch protection for VHE host apps
This patch enables branch protection for EL0 host apps that are build
with the test infrastructure. This patch also edits the VM memory
sharing tests to add the BTI instruction in tests where executable
memory is shared/lent. This is because the test VMs use the same page
table library as the hypervisor and the GP bit gets set, when it is
compiled branch protection enabled.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: Ib8ba887fe386c8e47e3e65487255a14665a74bc4