refactor(test): normal world EL0 test cleanup
This patch refactors source_set's and EL0 service related files to keep
it consistent with other tests. Until now, EL0 service infrastructure
was entirely isolated in the el0_partitions test directory. This patch
updates the build files so that the source sets live in test/hftest
directory and the foundation of a normal world EL0 partition
(nwd_el0_service.c) also lives in the hftest directory. This is also
in preparation to merge S-EL0 partition tests into the
ffa_secure_partitions test folder.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: Idf31cd807823b2613954b71e2c60bf09fc0e8d81