fix(ff-a): clear memory bit use in mem share from NWd to SWd
Requesting the SPMC to clear memory at the memory share call would not
work for an operation between the SWd and the NWd.
This was because the SPMC was not using the NS bit, when mapping the
memory region into its own address space before clearing the memory.
This patch makes SPMC map the memory of a memory send operation as NS,
if the lender is from the NWd.

This is a temporary fix, to avoid erroneous memory accesses from the
SPMC when clearing a memory region.
For a proper fix, the SPMC needs to keep track of the security state for
the memory region.

Change-Id: I0d640761d0f6b85ba20049f303d2b43c6e0a7135
Signed-off-by: J-Alves <joao.alves@arm.com>