feat(ff-a): handle forwarded FFA_VERSION from SPMD
The SPMD must forward calls to FFA_VERSION to the SPMC so that the
ff-a version of the callee can be saved for later use. Since the
return value of FFA_VERSION is not wrapped in a FF-A call we must
forward the call using a direct message request from the SPMD to
the SPMC. W2 of this request is used to identify the function
that should handle the message. Update the spmd_handler function
to choose the appropriate function to handle the message based on
W2 for both PSCI and FFA_VERSION.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: Ibf55f33130242303dd19f500f55abed5f06dbbef