fix(hftest): MP tests and S-EL0 partition
Introducing an S-EL0 partition into the 'ffa_secure_partitions' test has
broken the tests that run a MP setup.
Added function to fetch the information of the FF-A partitions in the
system. Using the UUID to determine the target of a test command, thus
helping to make an S-EL1 partition and an S-EL0 partition
interchangeable. Another S-EL1 partition was made the first to boot,
given it is an MP partition.
Power management tests validate that all SPs are running as expected.

Change-Id: If36378d93be4593bbfd1bedb7473c713e6426473
Signed-off-by: J-Alves <joao.alves@arm.com>