feat(gicv3): pass number of GICR frames from build env
Currently the amount of memory allocated for GICR frames is determined
by the number of supported CPUs. However, there can be more frames in
GIC redistributor than the number of CPUs. This can happen if some
CPUs are unavailable.

To get around this problem, pass the number of frames from build
environment explicitly. This is now specified in setting "gicr_frames"
that is specified in embedded.gni file. The size of allocated memory
is now 'GICR_FRAMES * GIC_REDIST_SIZE_PER_PE' bytes as a result.

NOTE: PLATFORMS MUST DEFINE 'gicr_frames' variable in their makefiles.

Change-Id: I616cdb36ae558491a1b8c58e75075af27fd6ac90
Signed-off-by: Bo Yan <byan@nvidia.com>
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>