refactor: factorize NWd VM shadow structure creation in SPMC
SPMC needs to keep track of NWd VMs for both notifications and RX/TX
mappings.
SPMC was creating the NWd VM structures on notification's
bitmap create, preventing the SPMC to map VM buffers before that
message.
Now the NWd VM structure creation is done when it's first needed,
breaking the dependency on notifications.

Change-Id: If737560258fb04b4cab5bfb19b03c530f6dcdee1
Signed-off-by: Federico Recanati <federico.recanati@arm.com>