fix: allocate hafnium manifest from bss section
Hafnium manifest structure has grown in size with addition of
FF-A fields. It is currently ~26KB. Init allocates the Hafnium
manifest from stack and leads to a stack corruption. CPU stacks
are 4KB in size. This change declares the Hafnium manifest from
BSS rather than stack.

Fixing the above exhibits a side effect with initrd begin and end
param values which must remain zero in context of the SPMC as no
ramdisk is provided.

Change-Id: I2b99e9879f9f90a699d395862692a0304e6a4487
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>