fix(mm): Fix stage 2 shareability
mm code currently marks any mapping in the stage 2 page tables as
non-shareable since it is "neutral". However, this is incorrect when
VCPUs migrate from one physical PE to another. If a VCPU has memory
marked as non-shareable in its stage 1 tables, and the stage 2 also had
the shareability marked as non-shareable, the resultant access would
also be non-shareable (D5-45 ARM ARM DDI0487G). A non-shareable access,
per the architecture, is not required to me made visible to other
observers, so when a vcpu makes a non-shareable but cacheable access on
one physical PE and then migrates to another physical PE (UP migratable
S-EL1 partitions for example), it would see incoherent values.
This patch changes the default to inner shareable in stage 2, to fix the
above problem.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: I5eed54d645ad4103f117a2d96e363c624af777e3