feat: add security state attribute definitions
FF-A v1.1 REL0 section 10.10.4.1 "Usage of NS bit" permits the SPMC to
provide the security state of a memory region through the memory
transaction descriptor attributes in a memory retrieve response.
Add the appropriate macro definitions.
Ensure a caller clears the security state attribute (NS-bit in
Table 10.18) in the memory transaction descriptor for a memory retrieve
request, donate, lend and share (section 10.10.4.1 bullet 1).

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: I9e93e37757ee4b7d7198d9f09494cbdee840de97