fix(arch): Add isb after sysreg write
ISB (context synchronization event) is required after write to system
registers. This is required for writes to CPTR_EL2 in this particular
change since arch_cpu_init is called during vcpu initialization in
vcpu_reset in the cold boot path. Following the initializtion call from
cpu_main, hafnium proceeds to vcpu_restore_all_and_run, which restores
simd registers. If an ISB is not inserted, the write to CPTR_EL2 is not
required to take effect architecturally and can cause the save/restore
of simd registers to trap in EL2.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: I263cc93aa6655996cc6e056def7904b3154e4839