feat(notifications): trigger NPI for framework notifications
Adding the respective field for framework notifications to the VM
structure. This is to initiate the support for framework notifications.

The function 'vm_are_global_notifications_pending' is used in
'plat_ffa_inject_notification_pending_interrupt' to determine if an NPI
should be triggered for pending global notifications.
Framework notifications are now included in such verification.

Change-Id: I5c68ae66c770eca57f645a2fb63a242be37c8ff5
Signed-off-by: J-Alves <joao.alves@arm.com>