ci: update submodules to use latest linux v5.10
Update driver, prebuilts and sources to linux kernel v5.10.

The reason is switching to a recent long term release, and more
importantly support proper testing of FF-A notifications feature in the
kernel module. The per cpu IRQ request for an SGI is only possible from
v5.9 because of additions in the arm64 support [1] [2].

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?h=v5.18-rc1&id=d3afc7f12987581eb0d1215b518d719fb9d762da
[2] https://lwn.net/Articles/820949/

Change-Id: I647f9ab8ea5da95aff7f4e9118d795c9680dd503
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>